feat(communication): reorganize tests, reexport libp2p types
